+++ Engine/Text.cpp
@@ -15,7 +15,7 @@ HRESULT Text::Init()
     assert(SUCCEEDED(dwrite_factory_->CreateTextFormat(L"Consolas", nullptr, DWRITE_FONT_WEIGHT_NORMAL, DWRITE_FONT_STYLE_NORMAL, DWRITE_FONT_STRETCH_NORMAL, 20, L"en-us", dwrite_text_format_.GetAddressOf())));
 
     // 텍스트 레이아웃 지정
-    assert(SUCCEEDED(dwrite_factory_->CreateTextLayout(default_text_.c_str(), default_text_.size(), dwrite_text_format_.Get(), g_client_rect.right, g_client_rect.bottom, dwrite_text_layout_.GetAddressOf())));
+    assert(SUCCEEDED(dwrite_factory_->CreateTextLayout(default_text_.c_str(), default_text_.size(), dwrite_text_format_.Get(), static_cast<float>(g_client_rect.right), static_cast<float>(g_client_rect.bottom), dwrite_text_layout_.GetAddressOf())));
 
     return TRUE;
 }
@@ -34,7 +34,7 @@ HRESULT Text::Render()
 
     if (Input::GetInstance()->GetKey(VK_LBUTTON) == KEY_STATE::NONE)
     {
-        Text::GetInstance()->Draw(Input::GetInstance()->GetPosition().x, Input::GetInstance()->GetPosition().y - 20, L"MOUSE" + point, { 1, 1, 1, 1 });
+        Text::GetInstance()->Draw((float)Input::GetInstance()->GetPosition().x, (float)Input::GetInstance()->GetPosition().y - 20, L"MOUSE" + point, { 1, 1, 1, 1 });
     }
 
     if (Input::GetInstance()->GetKey('Q') == KEY_STATE::HOLD)
@@ -127,7 +127,7 @@ HRESULT Text::SetSurface(Microsoft::WRL::ComPtr<IDXGISurface1> DXGISurface1)
 
 HRESULT Text::Draw(float x, float y, std::wstring text, D2D1_COLOR_F color)
 {
-    D2D1_RECT_F rect = { x, y, g_client_rect.right, g_client_rect.bottom };
+    D2D1_RECT_F rect = { x, y, (float)g_client_rect.right, (float)g_client_rect.bottom };
 
     d2d1_render_target_->BeginDraw();
